# modules/ordering/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order Processing",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "OrderingService"
    }
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/ordering": {
      "post": {
        "summary": "Start a new order",
        "operationId": "createOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbCreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderingpbCreateOrderRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/api/ordering/{id}": {
      "get": {
        "summary": "Get an order",
        "operationId": "getOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbGetOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Order"
        ]
      },
      "delete": {
        "summary": "Cancel an order",
        "operationId": "cancelOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbCancelOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Order"
        ]
      }
    }
  },
  "definitions": {
    "orderingpbCancelOrderResponse": {
      "type": "object"
    },
    "orderingpbCompleteOrderResponse": {
      "type": "object"
    },
    "orderingpbCreateOrderRequest": {
      "type": "object",
      "properties": {
        "idempotencyKey": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/orderingpbItem"
          }
        },
        "customerId": {
          "type": "string"
        },
        "paymentId": {
          "type": "string"
        }
      }
    },
    "orderingpbCreateOrderResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "accepted": {
          "type": "boolean",
          "title": "true if the order was already created due to idempotency"
        }
      }
    },
    "orderingpbGetOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/orderingpbOrder"
        }
      }
    },
    "orderingpbItem": {
      "type": "object",
      "properties": {
        "storeId": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "orderingpbOrder": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "customerId": {
          "type": "string"
        },
        "paymentId": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/orderingpbItem"
          }
        },
        "status": {
          "type": "string"
        }
      }
    },
    "orderingpbReadyOrderResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}